Move hook service to sentryapps
