+++ src/sentry/api/endpoints/project_servicehooks.py
@@ -13,7 +13,7 @@ from sentry.api.serializers import serialize
 from sentry.api.validators import ServiceHookValidator
 from sentry.constants import ObjectStatus
 from sentry.models.servicehook import ServiceHook
-from sentry.services.hybrid_cloud.hook import hook_service
+from sentry.sentry_apps.services.hook import hook_service
 
 
 @region_silo_endpoint
