+++ src/sentry/sentry_apps/installations.py
@@ -17,7 +17,7 @@ from sentry.models.integrations.sentry_app import SentryApp
 from sentry.models.integrations.sentry_app_installation import SentryAppInstallation
 from sentry.models.integrations.sentry_app_installation_token import SentryAppInstallationToken
 from sentry.models.user import User
-from sentry.services.hybrid_cloud.hook import hook_service
+from sentry.sentry_apps.services.hook import hook_service
 from sentry.tasks.sentry_apps import installation_webhook
 from sentry.utils import metrics
