+++ src/sentry/sentry_apps/services/hook/impl.py
@@ -5,8 +5,8 @@ from django.db import router, transaction
 from sentry import deletions
 from sentry.models.servicehook import ServiceHook
 from sentry.sentry_apps.apps import expand_events
-from sentry.services.hybrid_cloud.hook import HookService, RpcServiceHook
-from sentry.services.hybrid_cloud.hook.serial import serialize_service_hook
+from sentry.sentry_apps.services.hook import HookService, RpcServiceHook
+from sentry.sentry_apps.services.hook.serial import serialize_service_hook
 
 
 class DatabaseBackedHookService(HookService):
